requirements.txt
python-dotenv>=1.0.0,<2
markdown>=3.5.0,<4
google-generativeai>=0.7.0,<1

